requirements.txt
certifi==2019.3.9
chardet==3.0.4
get==2019.3.22
idna==2.8
post==2019.3.22
public==2019.3.22
query-string==2019.3.22
request==2019.3.22
requests==2.21.0
RPi.GPIO==0.6.5
urllib3==1.24.1